unsupported operand type(s) for / fixed

+++ images/users/src/services/learning_center.py
@@ -59,7 +59,8 @@ class LearningCenterService:
 
                     file_path = folder_path / file_name
                     file.save(file_path)
-                    Path(folder_path / old_file).unlink(missing_ok=True)
+                    if old_file:
+                        Path(folder_path / old_file).unlink(missing_ok=True)
 
         db.session.add(learning_center)
         db.session.commit()
